fixes: some refactoring
