--- app/equipment/armour_component.py
@@ -1,4 +1,0 @@
-class ArmourComponent:
-    def __init__(self, part: str, protection: int) -> None:
-        self.part = part
-        self.protection = protection
