--- app/equipment/potion.py
@@ -1,7 +1,0 @@
-from app.helpers.types import Effect
-
-
-class Potion:
-    def __init__(self, name: str, effect: Effect):
-        self.name = name
-        self.effect = effect
